+++ src/pages/Checkout/CheckoutPage.vue
@@ -213,6 +213,13 @@
 				@credit-removed="handleCreditRemoved"
 				@promo-opt-out-lightbox-closed="handleCancelPromoOptOut"
 			/>
+			<campaign-join-team-form
+				v-if="showTeamForm"
+				:campaign-name="campaignPartnerName"
+				:team-id="teamId"
+				:promo-id="promoFundId"
+				@team-process-complete="handleTeamJoinProcess"
+			/>
 
 			<div v-if="emptyBasket" class="empty-basket tw-relative tw-mx-auto" data-testid="empty-basket">
 				<div class="checkout-header-empty tw-mb-4">
@@ -257,6 +264,7 @@
 </template>
 
 <script>
+import gql from 'graphql-tag';
 import _get from 'lodash/get';
 import _filter from 'lodash/filter';
 import numeral from 'numeral';
@@ -272,6 +280,7 @@ import setupBasketForUserMutation from '@/graphql/mutation/shopSetupBasketForUse
 import validatePreCheckoutMutation from '@/graphql/mutation/shopValidatePreCheckout.graphql';
 import validationErrorsFragment from '@/graphql/fragments/checkoutValidationErrors.graphql';
 import experimentVersionFragment from '@/graphql/fragments/experimentVersion.graphql';
+import updateLoanReservationTeam from '@/graphql/mutation/updateLoanReservationTeam.graphql';
 import checkoutUtils from '@/plugins/checkout-utils-mixin';
 import KivaCreditPayment from '@/components/Checkout/KivaCreditPayment';
 import OrderTotals from '@/components/Checkout/OrderTotals';
@@ -280,6 +289,7 @@ import BasketVerification from '@/components/Checkout/BasketVerification';
 import KivaCardRedemption from '@/components/Checkout/KivaCardRedemption';
 import KvLoadingOverlay from '@/components/Kv/KvLoadingOverlay';
 import CampaignVerificationForm from '@/components/CorporateCampaign/CampaignVerificationForm';
+import CampaignJoinTeamForm from '@/components/CorporateCampaign/CampaignJoinTeamForm';
 import CheckoutHolidayPromo from '@/components/Checkout/CheckoutHolidayPromo';
 import CheckoutDropInPaymentWrapper from '@/components/Checkout/CheckoutDropInPaymentWrapper';
 import RandomLoanSelector from '@/components/RandomLoanSelector/RandomLoanSelector';
@@ -294,6 +304,21 @@ import { isLoanFundraising } from '@/util/loanUtils';
 import MatchedLoansLightbox from '@/components/Checkout/MatchedLoansLightbox';
 import KvPageContainer from '~/@kiva/kv-components/vue/KvPageContainer';
 import KvButton from '~/@kiva/kv-components/vue/KvButton';
+
+// Query to gather user Teams
+const myTeamsQuery = gql`query myTeamsQuery {
+	my {
+		lender {
+			id
+			teams(limit: 100) {
+				values {
+					id
+					name
+				}
+			}
+		}
+	}
+}`;
 
 export default {
 	name: 'CheckoutPage',
@@ -313,7 +338,8 @@ export default {
 		RandomLoanSelector,
 		VerifyRemovePromoCredit,
 		UpsellModule,
-		MatchedLoansLightbox
+		MatchedLoansLightbox,
+		CampaignJoinTeamForm
 	},
 	inject: ['apollo', 'cookieStore', 'kvAuth0'],
 	mixins: [
@@ -358,6 +384,8 @@ export default {
 			showUpsellModule: true,
 			requireDepositsMatchedLoans: false,
 			showMatchedLoansLightbox: false,
+			showTeamForm: false,
+			teamJoinStatus: null,
 			enableUpsellsCopy: false,
 		};
 	},
@@ -503,6 +531,13 @@ export default {
 				);
 			}
 		}
+		const matchedLoansWithCredit = this.loans?.filter(loan => {
+			const hasCredits = loan.creditsUsed?.length > 0;
+			const isMatchedLoan = loan.loan?.matchingText;
+			return hasCredits && isMatchedLoan;
+		});
+		this.showMatchedLoanKivaCredit = matchedLoansWithCredit.length > 0;
+		this.matchedText = matchedLoansWithCredit[0]?.loan?.matchingText ?? '';
 	},
 	mounted() {
 		// update current time every second for reactivity
@@ -529,6 +564,8 @@ export default {
 		this.handleToast();
 		this.getPromoInformationFromBasket();
 		this.getUpsellModuleData();
+		// show join-team form for specified scenario
+		this.handleTeamForm();
 	},
 	computed: {
 		isUpsellUnder100() {
@@ -621,7 +658,10 @@ export default {
 			return appliedCreditsPromoFunds[0] || null;
 		},
 		campaignPartnerName() {
-			return this.promoData?.promoFund?.displayName;
+			return this.promoData?.promoFund?.displayName ?? null;
+		},
+		teamId() {
+			return this.promoData?.promoGroup?.teamId ?? null;
 		},
 		externalFormId() {
 			return this.promoData?.managedAccount?.formId ?? null;
@@ -918,11 +958,60 @@ export default {
 				this.$kvTrackEvent('Lending', 'Add-to-Basket', 'Failed to add loan. Please try again.');
 				Sentry.captureException(error);
 			});
-		}
+		},
+		handleTeamJoinProcess(payload) {
+			this.teamJoinStatus = payload.join;
+			this.$kvTrackEvent(
+				'ManagedLendingCampaign',
+				'modal-join-team-process',
+				`${this.teamJoinStatus} team`
+			);
+			this.fetchMyTeams();
+		},
+		fetchMyTeams() {
+			this.apollo.query({
+				fetchPolicy: 'network-only',
+				query: myTeamsQuery
+			}).then(({ data }) => {
+				this.myTeams = data.my?.lender?.teams?.values ?? [];
+				if (this.teamJoinStatus !== 'declined') {
+					this.addTeamToLoans();
+				}
+			});
+		},
+		addTeamToLoans() {
+			if (this.loans.length && this.teamId) {
+				const loans = [];
+				// TODO Collect these promises and refresh basket once complete
+				this.loans.forEach((loan, index) => {
+					loans[index] = this.apollo.mutate({
+						mutation: updateLoanReservationTeam,
+						variables: {
+							teamId: this.teamId,
+							loanid: loan.id
+						}
+					});
+				});
+				Promise.all(loans).then(() => {
+					this.updateBasketState();
+				});
+			}
+		},
 	},
 	destroyed() {
 		clearInterval(this.currentTimeInterval);
 	},
+	handleTeamForm() {
+		if (
+			this.loans.length
+			&& this.isActivelyLoggedIn
+			&& this.teamId
+			&& !this.teamJoinStatus
+		) {
+			// check for team join optionality
+			this.showTeamForm = true;
+		}
+	}
 };
 </script>